+++ packages/core/src/columns/json/enum.ts
@@ -1,6 +1,7 @@
 import { constructType, JSONType, toCode } from './typeBase';
 import { singleQuoteArray } from '../../utils';
 
+// JSON enum type that consists of strings
 export interface JSONEnum<
   U extends string = string,
   T extends [U, ...U[]] = [U],
@@ -9,6 +10,7 @@ export interface JSONEnum<
   options: T;
 }
 
+// convert array to JSON enum type
 export const arrayToEnum = <U extends string, T extends [U, ...U[]]>(
   items: T,
 ) => {
@@ -19,6 +21,7 @@ export const arrayToEnum = <U extends string, T extends [U, ...U[]]>(
   return obj;
 };
 
+// JSON enum type constructor
 export const enumType = <U extends string, T extends [U, ...U[]]>(
   options: T,
 ) => {