+++ packages/core/src/columns/json/instanceOf.ts
@@ -1,5 +1,6 @@
 import { constructType, JSONType, toCode } from './typeBase';
 
+// JSON type to validate if object is instance of a class
 export interface JSONInstanceOf<T extends Class>
   extends JSONType<T, 'instanceOf'> {
   class: T;
@@ -8,6 +9,7 @@ export interface JSONInstanceOf<T extends Class>
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type Class = new (...args: any[]) => any;
 
+// constructor for instanceOf type
 export const instanceOf = <T extends Class>(cls: T) => {
   return constructType<JSONInstanceOf<T>>({
     dataType: 'instanceOf',