+++ packages/core/src/columns/json/literal.ts
@@ -1,11 +1,13 @@
 import { constructType, JSONType, Primitive, toCode } from './typeBase';
 import { singleQuote } from '../../utils';
 
+// JSON literal type
 export interface JSONLiteral<T extends Primitive>
   extends JSONType<T, 'literal'> {
   value: Primitive;
 }
 
+// constructor of JSON literal type
 export const literal = <T extends Primitive>(value: T) =>
   constructType<JSONLiteral<T>>({
     dataType: 'literal',