+++ packages/orm/src/table.ts
@@ -17,10 +17,15 @@ import {
 import { MapRelations, Relation, RelationThunks } from './relations/relations';
 import { OrchidORM } from './orm';
 
+// type of table class itself
 export type TableClass<T extends Table = Table> = new () => T;
 
+// object with table classes, used on orchidORM() for setting tables
 export type TableClasses = Record<string, TableClass>;
 
+// convert table instance type to queryable interface
+// processes relations to a type that's understandable by `pqb`
+// add ORM table specific metadata like `definedAt`, `db`, `filePath`
 export type TableToDb<T extends Table> = Db<
   T['table'],
   T['columns']['shape'],
@@ -38,36 +43,50 @@ export type TableToDb<T extends Table> = Db<
   T['columnTypes']
 > & { definedAs: string; db: OrchidORM; filePath: string; name: string };
 
+// convert a table class type into queryable interface
+// add relation methods
 export type DbTable<T extends TableClass> = TableToDb<InstanceType<T>> &
   Omit<MapRelations<InstanceType<T>>, keyof Query>;
 
-type TableConfig = {
+// `columns` property of table has a shape and an output type of the columns
+type ColumnsConfig = {
   shape: ColumnsShape;
   type: unknown;
 };
 
+// callback with a query of relation, to use as a default scope
 type ScopeFn<Related extends TableClass, Scope extends Query> = (
   q: DbTable<Related>,
 ) => Scope;
 
+// type of table instance created by a table class
 export type Table = {
   table: string;
-  columns: TableConfig;
+  columns: ColumnsConfig;
   schema?: string;
   columnTypes: ColumnTypesBase;
   noPrimaryKey?: boolean;
   filePath: string;
 };
 
+// base table constructor
 export const createBaseTable = <CT extends ColumnTypesBase>(
   {
     columnTypes,
     snakeCase,
     filePath,
+    nowSQL,
   }: {
+    // concrete column types or a callback for overriding standard column types
+    // this types will be used in tables to define their columns
     columnTypes?: CT | ((t: DefaultColumnTypes) => CT);
+    // when set to true, all columns will be translated to `snake_case` when querying database
     snakeCase?: boolean;
+    // if for some unknown reason you see error that file path for a table can't be guessed automatically,
+    // provide it manually via `filePath`
     filePath?: string;
+    // if `now()` for some reason doesn't suite your timestamps, provide a custom SQL for it
+    nowSQL?: string;
   } = { columnTypes: defaultColumnTypes as unknown as CT },
 ) => {
   const ct =
@@ -86,6 +105,7 @@ export const createBaseTable = <CT extends ColumnTypesBase>(
     ct as ColumnTypesBase extends CT ? DefaultColumnTypes : CT,
     filePath,
     snakeCase,
+    nowSQL,
   );
 };
 
@@ -93,12 +113,14 @@ const create = <CT extends ColumnTypesBase>(
   columnTypes: CT,
   filePath: string,
   snakeCase?: boolean,
+  nowSQL?: string,
 ) => {
   const base = class BaseTable {
     static filePath = filePath;
+    static nowSQL = nowSQL;
 
     table!: string;
-    columns!: TableConfig;
+    columns!: ColumnsConfig;
     schema?: string;
     noPrimaryKey?: boolean;
     snakeCase = snakeCase;
@@ -126,7 +148,7 @@ const create = <CT extends ColumnTypesBase>(
       (columnTypes as { [snakeCaseKey]?: boolean })[snakeCaseKey] =
         this.snakeCase;
 
-      const shape = getColumnTypes(columnTypes, fn);
+      const shape = getColumnTypes(columnTypes, fn, nowSQL);
 
       if (this.snakeCase) {
         for (const key in shape) {