+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # orchid-core
 
+## 0.4.6
+
+### Patch Changes
+
+- Allow to customize timestamp default value with `nowSQL` option of base table
+
 ## 0.4.5
 
 ### Patch Changes