+++ packages/qb/myqb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # myqb
 
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies
+  - orchid-core@0.4.6
+
 ## 0.2.8
 
 ### Patch Changes